+++ ui/CountlyFeedbackDemo.Wpf/App.xaml.cs
@@ -1,0 +1,6 @@
+namespace CountlyFeedbackDemo.Wpf
+{
+    public partial class App : System.Windows.Application
+    {
+    }
+}